requirements.txt
python-levenshtein
yt-dlp
pandas
regex
python-vlc